// tests-unit/stream-async.gen.ts
export interface StreamTestClient {
  streamMessages(req: StreamMessagesRequest, options?: WebrpcOptions): WebrpcStream<StreamMessagesResponse>

  streamTicks(options?: WebrpcOptions): WebrpcStream<StreamTicksResponse>
}


//
// Schema types
//

export interface Message {
  id: number
  text: string
}

export interface StreamMessagesRequest {
  roomId: string
}

export interface StreamMessagesResponse {
  message: Message
}

export interface StreamTicksRequest {
}

export interface StreamTicksResponse {
  tick: number
}



//
// Client
//

export class StreamTest implements StreamTestClient {
  protected hostname: string
  protected fetch: Fetch
  protected path = '/rpc/StreamTest/'

  constructor(hostname: string, fetch: Fetch) {
    this.hostname = hostname.replace(/\/*$/, '')
    this.fetch = (input: RequestInfo, init?: RequestInit) => fetch(input, init)
  }

  private url(name: string): string {
    return this.hostname + this.path + name
  }
  
  queryKey = {
    streamMessages: (req: StreamMessagesRequest) => ['StreamTest', 'streamMessages', req] as const,
    streamTicks: () => ['StreamTest', 'streamTicks'] as const,
  }

  streamMessages = (req: StreamMessagesRequest, options?: WebrpcOptions): WebrpcStream<StreamMessagesResponse> => {
    return webrpcAsyncIterable<StreamMessagesResponse>(() => this.fetch(this.url('StreamMessages'),
      createHttpRequest(JsonEncode(req), options?.headers, options?.signal)
    ), 'StreamMessagesResponse', options?.signal)
  }
  streamTicks = (options?: WebrpcOptions): WebrpcStream<StreamTicksResponse> => {
    return webrpcAsyncIterable<StreamTicksResponse>(() => this.fetch(this.url('StreamTicks'),
      createHttpRequest('{}', options?.headers, options?.signal)
    ), 'StreamTicksResponse', options?.signal)
  }
}
// ...
export type Fetch = (input: RequestInfo, init?: RequestInit) => Promise<Response>


export interface WebrpcOptions {
  headers?: HeadersInit;
  signal?: AbortSignal;
}

export type WebrpcStream<T> = AsyncIterable<T>
// ...
type WebrpcErrorParams = { name?: string, code?: number, message?: string, status?: number, cause?: string }

export class WebrpcError extends Error {
  code: number
  status: number

  constructor(error: WebrpcErrorParams = {}) {
    super(error.message)
    this.name = error.name || 'WebrpcEndpointError'
    this.code = typeof error.code === 'number' ? error.code : 0
    this.message = error.message || `endpoint error`
    this.status = typeof error.status === 'number' ? error.status : 400
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, WebrpcError.prototype)
  }

  static new(payload: any): WebrpcError {
    return new this({ message: payload.message, code: payload.code, status: payload.status, cause: payload.cause })
  }
}


export class WebrpcEndpointError extends WebrpcError {
  constructor(error: WebrpcErrorParams = {}) {
    super(error)
    this.name = error.name || 'WebrpcEndpoint'
    this.code = typeof error.code === 'number' ? error.code : 0
    this.message = error.message || `endpoint error`
    this.status = typeof error.status === 'number' ? error.status : 400
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, WebrpcEndpointError.prototype)
  }
}
// ...
export class WebrpcStreamLostError extends WebrpcError {
  constructor(error: WebrpcErrorParams = {}) {
    super(error)
    this.name = error.name || 'WebrpcStreamLost'
    this.code = typeof error.code === 'number' ? error.code : -9
    this.message = error.message || `stream lost`
    this.status = typeof error.status === 'number' ? error.status : 400
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, WebrpcStreamLostError.prototype)
  }
}
// ...
export class StreamBoomError extends WebrpcError {
  constructor(error: WebrpcErrorParams = {}) {
    super(error)
    this.name = error.name || 'StreamBoom'
    this.code = typeof error.code === 'number' ? error.code : 100
    this.message = error.message || `stream boom`
    this.status = typeof error.status === 'number' ? error.status : 500
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, StreamBoomError.prototype)
  }
}